+++ extmod/machine_i2s.c
@@ -93,30 +93,8 @@ enum {
     ARG_ibuf,
 };
 
-#if MICROPY_PY_MACHINE_I2S_RING_BUF
-
-typedef struct _ring_buf_t {
-    uint8_t *buffer;
-    size_t head;
-    size_t tail;
-    size_t size;
-} ring_buf_t;
-
-typedef struct _non_blocking_descriptor_t {
-    mp_buffer_info_t appbuf;
-    uint32_t index;
-    bool copy_in_progress;
-} non_blocking_descriptor_t;
-
-static void ringbuf_init(ring_buf_t *rbuf, uint8_t *buffer, size_t size);
-static bool ringbuf_push(ring_buf_t *rbuf, uint8_t data);
-static bool ringbuf_pop(ring_buf_t *rbuf, uint8_t *data);
-static size_t ringbuf_available_data(ring_buf_t *rbuf);
-static size_t ringbuf_available_space(ring_buf_t *rbuf);
-static void fill_appbuf_from_ringbuf_non_blocking(machine_i2s_obj_t *self);
-static void copy_appbuf_to_ringbuf_non_blocking(machine_i2s_obj_t *self);
-
-#endif // MICROPY_PY_MACHINE_I2S_RING_BUF
+//prepare common/shared audio type and function definitions and required port implementations
+#include "extmod/machine_audio_common.h"
 
 // The port must provide implementations of these low-level I2S functions.
 static void mp_machine_i2s_init_helper(machine_i2s_obj_t *self, mp_arg_val_t *args);
@@ -124,221 +102,12 @@ static machine_i2s_obj_t *mp_machine_i2s_make_new_instance(mp_int_t i2s_id);
 static void mp_machine_i2s_deinit(machine_i2s_obj_t *self);
 static void mp_machine_i2s_irq_update(machine_i2s_obj_t *self);
 
-// The port provides implementations of the above in this file.
+// The port provides implementations of the above via include file.
+// The port must provide the implementations of the above header files.
 #include MICROPY_PY_MACHINE_I2S_INCLUDEFILE
 
 #if MICROPY_PY_MACHINE_I2S_RING_BUF
-
-// Ring Buffer
-// Thread safe when used with these constraints:
-// - Single Producer, Single Consumer
-// - Sequential atomic operations
-// One byte of capacity is used to detect buffer empty/full
-
-static void ringbuf_init(ring_buf_t *rbuf, uint8_t *buffer, size_t size) {
-    rbuf->buffer = buffer;
-    rbuf->size = size;
-    rbuf->head = 0;
-    rbuf->tail = 0;
-}
-
-static bool ringbuf_push(ring_buf_t *rbuf, uint8_t data) {
-    size_t next_tail = (rbuf->tail + 1) % rbuf->size;
-
-    if (next_tail != rbuf->head) {
-        rbuf->buffer[rbuf->tail] = data;
-        rbuf->tail = next_tail;
-        return true;
-    }
-
-    // full
-    return false;
-}
-
-static bool ringbuf_pop(ring_buf_t *rbuf, uint8_t *data) {
-    if (rbuf->head == rbuf->tail) {
-        // empty
-        return false;
-    }
-
-    *data = rbuf->buffer[rbuf->head];
-    rbuf->head = (rbuf->head + 1) % rbuf->size;
-    return true;
-}
-
-static bool ringbuf_is_empty(ring_buf_t *rbuf) {
-    return rbuf->head == rbuf->tail;
-}
-
-static bool ringbuf_is_full(ring_buf_t *rbuf) {
-    return ((rbuf->tail + 1) % rbuf->size) == rbuf->head;
-}
-
-static size_t ringbuf_available_data(ring_buf_t *rbuf) {
-    return (rbuf->tail - rbuf->head + rbuf->size) % rbuf->size;
-}
-
-static size_t ringbuf_available_space(ring_buf_t *rbuf) {
-    return rbuf->size - ringbuf_available_data(rbuf) - 1;
-}
-
-static uint32_t fill_appbuf_from_ringbuf(machine_i2s_obj_t *self, mp_buffer_info_t *appbuf) {
-
-    // copy audio samples from the ring buffer to the app buffer
-    // loop, copying samples until the app buffer is filled
-    // For asyncio mode, the loop will make an early exit if the ring buffer becomes empty
-    // Example:
-    //   a MicroPython I2S object is configured for 16-bit mono (2 bytes per audio sample).
-    //   For every frame coming from the ring buffer (8 bytes), 2 bytes are "cherry picked" and
-    //   copied to the supplied app buffer.
-    //   Thus, for every 1 byte copied to the app buffer, 4 bytes are read from the ring buffer.
-    //   If a 8kB app buffer is supplied, 32kB of audio samples is read from the ring buffer.
-
-    uint32_t num_bytes_copied_to_appbuf = 0;
-    uint8_t *app_p = (uint8_t *)appbuf->buf;
-    uint8_t appbuf_sample_size_in_bytes = (self->bits == 16? 2 : 4) * (self->format == STEREO ? 2: 1);
-    uint32_t num_bytes_needed_from_ringbuf = appbuf->len * (I2S_RX_FRAME_SIZE_IN_BYTES / appbuf_sample_size_in_bytes);
-    uint8_t discard_byte;
-    while (num_bytes_needed_from_ringbuf) {
-
-        uint8_t f_index = get_frame_mapping_index(self->bits, self->format);
-
-        for (uint8_t i = 0; i < I2S_RX_FRAME_SIZE_IN_BYTES; i++) {
-            int8_t r_to_a_mapping = i2s_frame_map[f_index][i];
-            if (r_to_a_mapping != -1) {
-                if (self->io_mode == BLOCKING) {
-                    // poll the ringbuf until a sample becomes available,  copy into appbuf using the mapping transform
-                    while (ringbuf_pop(&self->ring_buffer, app_p + r_to_a_mapping) == false) {
-                        ;
-                    }
-                    num_bytes_copied_to_appbuf++;
-                } else if (self->io_mode == ASYNCIO) {
-                    if (ringbuf_pop(&self->ring_buffer, app_p + r_to_a_mapping) == false) {
-                        // ring buffer is empty, exit
-                        goto exit;
-                    } else {
-                        num_bytes_copied_to_appbuf++;
-                    }
-                } else {
-                    return 0;  // should never get here (non-blocking mode does not use this function)
-                }
-            } else { // r_a_mapping == -1
-                // discard unused byte from ring buffer
-                if (self->io_mode == BLOCKING) {
-                    // poll the ringbuf until a sample becomes available
-                    while (ringbuf_pop(&self->ring_buffer, &discard_byte) == false) {
-                        ;
-                    }
-                } else if (self->io_mode == ASYNCIO) {
-                    if (ringbuf_pop(&self->ring_buffer, &discard_byte) == false) {
-                        // ring buffer is empty, exit
-                        goto exit;
-                    }
-                } else {
-                    return 0;  // should never get here (non-blocking mode does not use this function)
-                }
-            }
-            num_bytes_needed_from_ringbuf--;
-        }
-        app_p += appbuf_sample_size_in_bytes;
-    }
-exit:
-    return num_bytes_copied_to_appbuf;
-}
-
-// function is used in IRQ context
-static void fill_appbuf_from_ringbuf_non_blocking(machine_i2s_obj_t *self) {
-
-    // attempt to copy a block of audio samples from the ring buffer to the supplied app buffer.
-    // audio samples will be formatted as part of the copy operation
-
-    uint32_t num_bytes_copied_to_appbuf = 0;
-    uint8_t *app_p = &(((uint8_t *)self->non_blocking_descriptor.appbuf.buf)[self->non_blocking_descriptor.index]);
-
-    uint8_t appbuf_sample_size_in_bytes = (self->bits == 16? 2 : 4) * (self->format == STEREO ? 2: 1);
-    uint32_t num_bytes_remaining_to_copy_to_appbuf = self->non_blocking_descriptor.appbuf.len - self->non_blocking_descriptor.index;
-    uint32_t num_bytes_remaining_to_copy_from_ring_buffer = num_bytes_remaining_to_copy_to_appbuf *
-        (I2S_RX_FRAME_SIZE_IN_BYTES / appbuf_sample_size_in_bytes);
-    uint32_t num_bytes_needed_from_ringbuf = MIN(SIZEOF_NON_BLOCKING_COPY_IN_BYTES, num_bytes_remaining_to_copy_from_ring_buffer);
-    uint8_t discard_byte;
-    if (ringbuf_available_data(&self->ring_buffer) >= num_bytes_needed_from_ringbuf) {
-        while (num_bytes_needed_from_ringbuf) {
-
-            uint8_t f_index = get_frame_mapping_index(self->bits, self->format);
-
-            for (uint8_t i = 0; i < I2S_RX_FRAME_SIZE_IN_BYTES; i++) {
-                int8_t r_to_a_mapping = i2s_frame_map[f_index][i];
-                if (r_to_a_mapping != -1) {
-                    ringbuf_pop(&self->ring_buffer, app_p + r_to_a_mapping);
-                    num_bytes_copied_to_appbuf++;
-                } else { // r_a_mapping == -1
-                    // discard unused byte from ring buffer
-                    ringbuf_pop(&self->ring_buffer, &discard_byte);
-                }
-                num_bytes_needed_from_ringbuf--;
-            }
-            app_p += appbuf_sample_size_in_bytes;
-        }
-        self->non_blocking_descriptor.index += num_bytes_copied_to_appbuf;
-
-        if (self->non_blocking_descriptor.index >= self->non_blocking_descriptor.appbuf.len) {
-            self->non_blocking_descriptor.copy_in_progress = false;
-            mp_sched_schedule(self->callback_for_non_blocking, MP_OBJ_FROM_PTR(self));
-        }
-    }
-}
-
-static uint32_t copy_appbuf_to_ringbuf(machine_i2s_obj_t *self, mp_buffer_info_t *appbuf) {
-
-    // copy audio samples from the app buffer to the ring buffer
-    // loop, reading samples until the app buffer is emptied
-    // for asyncio mode, the loop will make an early exit if the ring buffer becomes full
-
-    uint32_t a_index = 0;
-
-    while (a_index < appbuf->len) {
-        if (self->io_mode == BLOCKING) {
-            // copy a byte to the ringbuf when space becomes available
-            while (ringbuf_push(&self->ring_buffer, ((uint8_t *)appbuf->buf)[a_index]) == false) {
-                ;
-            }
-            a_index++;
-        } else if (self->io_mode == ASYNCIO) {
-            if (ringbuf_push(&self->ring_buffer, ((uint8_t *)appbuf->buf)[a_index]) == false) {
-                // ring buffer is full, exit
-                break;
-            } else {
-                a_index++;
-            }
-        } else {
-            return 0;  // should never get here (non-blocking mode does not use this function)
-        }
-    }
-
-    return a_index;
-}
-
-// function is used in IRQ context
-static void copy_appbuf_to_ringbuf_non_blocking(machine_i2s_obj_t *self) {
-
-    // copy audio samples from app buffer into ring buffer
-    uint32_t num_bytes_remaining_to_copy = self->non_blocking_descriptor.appbuf.len - self->non_blocking_descriptor.index;
-    uint32_t num_bytes_to_copy = MIN(SIZEOF_NON_BLOCKING_COPY_IN_BYTES, num_bytes_remaining_to_copy);
-
-    if (ringbuf_available_space(&self->ring_buffer) >= num_bytes_to_copy) {
-        for (uint32_t i = 0; i < num_bytes_to_copy; i++) {
-            ringbuf_push(&self->ring_buffer,
-                ((uint8_t *)self->non_blocking_descriptor.appbuf.buf)[self->non_blocking_descriptor.index + i]);
-        }
-
-        self->non_blocking_descriptor.index += num_bytes_to_copy;
-        if (self->non_blocking_descriptor.index >= self->non_blocking_descriptor.appbuf.len) {
-            self->non_blocking_descriptor.copy_in_progress = false;
-            mp_sched_schedule(self->callback_for_non_blocking, MP_OBJ_FROM_PTR(self));
-        }
-    }
-}
-
+#include "extmod/machine_audio_common.c"
 #endif // MICROPY_PY_MACHINE_I2S_RING_BUF
 
 MP_NOINLINE static void machine_i2s_init_helper(machine_i2s_obj_t *self, size_t n_pos_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
